구현한 API를 테스트컨트롤러에서 분리
